+++ src/VaultSharp/V1/AuthMethods/Custom/CustomAuthMethodInfo.cs
@@ -23,25 +23,28 @@ namespace VaultSharp.V1.AuthMethods.Custom
         }
 
         /// <summary>
-        /// Gets the authentication token asynchronous delegate.
+        /// Gets the CustomAuthMethodInfo asynchronous delegate which includes the token and token info.
         /// </summary>
-        /// <value>
-        /// The authentication token asynchronous delegate.
-        /// </value>
-        public Func<Task<string>> AuthenticationTokenAsyncDelegate { get; }
+        public Func<Task<CustomAuthMethodInfo>> CustomAuthMethodInfoAsyncDelegate { get; }
+
+        /// <summary>
+        /// Use this version in the delegate to return a CustomAuthMethodInfo.
+        /// </summary>
+        public CustomAuthMethodInfo()
+        { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="CustomAuthMethodInfo"/> class.
         /// </summary>
-        /// <param name="type">The type of the unknown authentication backend type not supported by this library yet. e.g. liveid, facebook etc.</param>
-        /// <param name="authenticationTokenAsyncDelegate">The authentication token asynchronous delegate.</param>
-        public CustomAuthMethodInfo(string type, Func<Task<string>> authenticationTokenAsyncDelegate)
+        /// <param name="type">The type of the unknown authentication backend type not supported by this library yet. But supported by the Vault Server.</param>
+        /// <param name="customAuthMethodInfoAsyncDelegate">The authentication token asynchronous delegate.</param>
+        public CustomAuthMethodInfo(string type, Func<Task<CustomAuthMethodInfo>> customAuthMethodInfoAsyncDelegate)
         {
             Checker.NotNull(type, "type");
-            Checker.NotNull(authenticationTokenAsyncDelegate, "authenticationTokenAsyncDelegate");
+            Checker.NotNull(customAuthMethodInfoAsyncDelegate, nameof(customAuthMethodInfoAsyncDelegate));
 
             _lazyAuthMethodType = new Lazy<AuthMethodType>(() => new AuthMethodType(type));
-            AuthenticationTokenAsyncDelegate = authenticationTokenAsyncDelegate;
+            CustomAuthMethodInfoAsyncDelegate = customAuthMethodInfoAsyncDelegate;
         }
     }
 }